requirements-kg.txt
aerich==0.9.1
aiohappyeyeballs==2.6.1
aiokafka==0.12.0
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncclick==8.1.8.0
atlastk==0.13.5
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
dictdiffer==0.9.0
dmPython==2.5.22
dmSQLAlchemy==2.0.2
dnspython==2.7.0
elasticsearch==7.17.0
email_validator==2.2.0
fastapi==0.116.1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
happybase==1.2.0
idna==3.10
iso8601==2.1.0
kafka-python==2.2.15
loguru==0.7.3
multidict==6.6.3
neo4j==5.28.1
packaging==25.0
passlib==1.7.4
ply==3.11
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pypika-tortoise==0.6.1
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
requests==2.32.4
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
thriftpy2==0.5.3
tortoise==0.1.1
tortoise-orm==0.25.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.26.20
uvicorn==0.35.0
wheel==0.45.1
yarl==1.20.1
aiohttp